Export SimpleFormView
