+++ packages/ra-ui-materialui/src/form/index.tsx
@@ -6,17 +6,24 @@ import SimpleFormIterator, {
 import TabbedFormTabs from './TabbedFormTabs';
 import Toolbar, { ToolbarProps } from './Toolbar';
 import getFormInitialValues from './getFormInitialValues';
+import { SimpleFormView, SimpleFormViewProps } from './SimpleFormView';
 export * from './FormTabHeader';
 
 export * from './TabbedForm';
 export * from './FormTab';
 export * from './FormTabHeader';
 
-export type { SimpleFormProps, SimpleFormIteratorProps, ToolbarProps };
+export type {
+    SimpleFormProps,
+    SimpleFormIteratorProps,
+    SimpleFormViewProps,
+    ToolbarProps,
+};
 
 export {
     FormInput,
     SimpleForm,
+    SimpleFormView,
     SimpleFormIterator,
     TabbedFormTabs,
     Toolbar,
